+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix a `metaTx` malleability bug in `CrossChainReceiverFactory`
   (contract is not deployed; no funds at risk) after a report in
   Immunefi bug 78645
+* Add `FLUXPOOL` action on Bnb
 
 ## 2026-09-03